Update README.md
Project Description
Cookie Cats is a hugely popular mobile puzzle game developed by Tactile Entertainment. It's a classic "connect three" style puzzle game where the player must connect tiles of the same color in order to clear the board and win the level. It also features singing cats. We're not kidding! Check out this short demo:
As players progress through the game they will encounter gates that force them to wait some time before they can progress or make an in-app purchase. In this project, we will analyze the result of an A/B test where the first gate in Cookie Cats was moved from level 30 to level 40. In particular, we will analyze the impact on player retention.

Project Tasks
1.	Of cats and cookies
2.	The AB-test data
3.	The distribution of game rounds
4.	Overall 1-day retention
5.	1-day retention by AB-group
6.	Should we be confident in the difference?
7.	Zooming in on the difference
8.	The probability of a difference
9.	7-day retention by AB-group
10.	Bootstrapping the difference again
11.	The conclusion